+++ src/main/java/kitchenpos/domain/JpaProductRepository.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
