+++ src/main/java/kitchenpos/product/README.md
@@ -1,0 +1,21 @@
+## 용어 사전
+
+### 상품 (Bounded Context)
+
+| 한글명 | 영문명 | 설명 |
+| --- | --- | --- |
+| 상품 | product | 판매할 음식의 종류를 의미한다. 메뉴는 하나 이상의 상품으로 구성된다. |
+| 상품 가격 | productPrice | 상품의 가격을 의미한다. |
+| 상품명 | productName | 상품의 이름을 의미한다. |
+| 비속어 | profanity | 욕설 또는 비속어로 손님들이 불편함을 느껴 상품명에 포함될 수 없는 단어를 의미한다. |
+
+## 모델링
+
+### 상품
+
+- `상품(product)`은 식별자, 이름, 가격을 가진다.<br/><br/>
+- `상품(product)`을 생성한다.
+    - `상품명(productName)`에는 `비속어(profanity)`가 포함되면 안된다.
+    - `상품 가격(productPrice)`은 0원 이상이어야 한다.
+- `상품 가격(productPrice)`을 변경할 수 있다.
+    - 상품의 가격이 변경될 때, `해당 상품을 포함하는 메뉴`의 `메뉴 상품(menuProduct)`의 가격 총합이 상품을 포함하는 메뉴의 가격보다 낮아지면 해당 메뉴를 숨긴다.
