+++ src/main/java/kitchenpos/product/infra/ProductPurgomalumClient.java
@@ -1,0 +1,26 @@
+package kitchenpos.product.infra;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import java.net.URI;
+
+@Component
+public class ProductPurgomalumClient implements PurgomalumClient {
+    private final RestTemplate restTemplate;
+
+    public ProductPurgomalumClient(final RestTemplateBuilder restTemplateBuilder) {
+        this.restTemplate = restTemplateBuilder.build();
+    }
+
+    @Override
+    public boolean containsProfanity(final String text) {
+        final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
+            .queryParam("text", text)
+            .build()
+            .toUri();
+        return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
+    }
+}
