+++ src/main/java/kitchenpos/product/infra/PurgomalumClient.java
@@ -1,0 +1,5 @@
+package kitchenpos.product.infra;
+
+public interface PurgomalumClient {
+    boolean containsProfanity(String text);
+}
