+++ src/test/java/kitchenpos/application/FakeProductPurgomalumClient.java
@@ -1,11 +1,11 @@
 package kitchenpos.application;
 
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.product.infra.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class FakePurgomalumClient implements PurgomalumClient {
+public class FakeProductPurgomalumClient implements PurgomalumClient {
     private static final List<String> profanities;
 
     static {
